DRF/cinema/requirements.txt
django
djangorestframework
requests
pillow
django-cors-headers